requirements.txt
absl-py==2.3.1
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alabaster==1.0.0
anndata==0.11.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
array-api-compat==1.12.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
cached-property==2.0.1
cell-gears==0.0.2
cell2sentence==1.1.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
chex==0.1.89
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.3
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==2.21.0
dcor==0.6
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docrep==0.3.2
docutils==0.21.2
et_xmlfile==2.0.0
etils==1.12.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.14
fastjsonschema==2.21.1
fastobo==0.13.0
filelock==3.18.0
flax==0.10.7
fonttools==4.58.5
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.6.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanize==4.12.3
idna==3.10
igraph==0.11.9
imagesize==1.4.1
importlib_resources==6.5.2
interegular==0.3.3
ipykernel==6.29.5
ipython==8.37.0
isoduration==20.11.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lark==1.2.2
legacy-api-wrap==1.4.1
leidenalg==0.10.2
lightning-utilities==0.15.2
llvmlite==0.44.0
lm-format-enforcer==0.9.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mizani==0.13.5
ml_collections==1.1.0
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
mudata==0.3.2
multidict==6.6.4
multipledispatch==1.0.0
multiprocess==0.70.16
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
numpyro==0.18.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
openai==1.93.0
openpyxl==3.1.5
opt_einsum==3.4.0
optax==0.2.5
orbax==0.1.7
orbax-checkpoint==0.11.18
outlines==0.0.34
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotnine==0.14.6
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
pronto==2.7.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytorch-lightning==1.9.5
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
ray==2.47.1
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.26.0
safetensors==0.6.2
scanpy==1.11.3
scgpt==0.2.4
scib==1.1.7
scikit-learn==1.7.1
scikit-misc==0.5.1
scipy==1.15.3
scvi-tools==0.20.3
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
session-info2==0.1.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
sympy==1.14.0
tensorstore==0.1.76
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tiktoken==0.6.0
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.2.1
toolz==1.0.0
torch==2.3.0+cu121
torch-geometric==2.6.1
torchaudio==2.3.0+cu121
torchdata==0.8.0
torchmetrics==1.7.3
torchtext==0.18.0
torchvision==0.18.0+cu121
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.0
treescope==0.1.9
triton==2.3.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.4.2
vllm_nccl_cu12==2.18.1.0.4.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xformers==0.0.26.post1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
